+++ data-migrations/2024-06-07-1437-seed-oauthClients.sql
@@ -1,8 +1,7 @@
-INSERT INTO `oauthClients`
-  (`name`, `redirectUris`, `grants`, `clientId`, `clientSecret`, `userId`)
-VALUES
-  (SELECT
-    'tester', 'http://localhost:3000/auth/callback', 'client_credentials, authorization_code',
-    '1234567890', UUID(), `users`.`id`
-  FROM `users`
-  WHERE `email` = 'org.admin@example.com');
+INSERT INTO oauthClients
+  (name, redirectUris, grants, clientId, clientSecret, userId)
+(SELECT
+  'tester', 'http://localhost:3000/auth/callback', 'client_credentials, authorization_code',
+  '1234567890', UUID(), `users`.`id`
+FROM users
+WHERE users.email = 'orgA.admin@example.com');