+++ src/context.ts
@@ -1,5 +1,6 @@
 import { Logger } from 'pino';
-import { DMPHubAPI } from './datasources/dmphub-api';
+import { DMPHubAPI } from './datasources/dmphubAPI';
+import { DMPToolAPI } from './datasources/dmptoolAPI';
 import { MySQLDataSource } from './datasources/mySQLDataSource';
 import { User } from './models/User';
 
@@ -10,6 +11,7 @@ export type MyContext = {
 
   dataSources: {
     dmphubAPIDataSource: DMPHubAPI;
+    dmptoolAPIDataSource: DMPToolAPI;
     sqlDataSource: MySQLDataSource;
   };
 }