+++ src/models/User.ts
@@ -10,7 +10,7 @@ import { logger, formatLogMessage } from '../logger';
 export enum UserRole {
   Researcher = 'RESEARCHER',
   Admin = 'ADMIN',
-  Super = 'SUPER_ADMIN',
+  SuperAdmin = 'SUPER_ADMIN',
 }
 
 export class User {
@@ -24,6 +24,9 @@ export class User {
   public surName?: string;
   public orcid?: string;
   public errors: string[];
+  public affiliationId: string;
+  public created: string;
+  public modified: string;
 
   // Initialize a new User
   constructor(options) {
@@ -35,6 +38,9 @@ export class User {
     this.givenName = options.givenName;
     this.surName = options.surName;
     this.orcid = options.orcid;
+    this.affiliationId = options.affiliationId;
+    this.created = options.created || new Date().toUTCString();
+    this.modified = options.modified || new Date().toUTCString();
     this.errors = [];
 
     this.cleanup();
@@ -101,9 +107,10 @@ export class User {
   }
 
   // Find the User by their Id
-  static async findById(userId: string): Promise<User | Falsey> {
+  static async findById(userId: string): Promise<User> {
     const mysql = MySQLDataSource.getInstance();
-    const sql = 'SELECT * FROM users WHERE id = ?';
+    const sql = 'SELECT id, email, givenName, surName, role, affiliationId, created, modified \
+                 FROM users WHERE id = ?';
     formatLogMessage(logger)?.debug(`User.findById: ${userId}`);
     try {
       const [rows] = await mysql.query(sql, [userId]);
@@ -115,9 +122,10 @@ export class User {
   }
 
   // Find the User by their email address
-  static async findByEmail(email: string): Promise<User | Falsey> {
+  static async findByEmail(email: string): Promise<User | null> {
     const mysql = MySQLDataSource.getInstance();
-    const sql = 'SELECT * from users where email = ?';
+    const sql = 'SELECT id, email, givenName, surName, role, affiliationId, created, modified \
+                from users where email = ?';
 
     formatLogMessage(logger)?.debug(`User.findByEmail: ${email}`);
     try {